requirements.txt
typer
rich
PyYAML
openai>=1.0
